CountDownLatches

CountDownLatches.CountDownLatch

CountDownLatches.await

CountDownLatches.count_down

CountDownLatches.get_count

CountDownLatches.CountDownLatch — Type

CountDownLatch(count::I) where {I<:Integer}

A CountDownLatch that starts at some initial non-negative count. This latch can be counted down and waited on.

CountDownLatches.await — Method

await(latch::CountDownLatch)

Waits until the latch has counted down to <= 0.

CountDownLatches.count_down — Method

count_down(latch::CountDownLatch)

Counts down the latch once. This may cause the count to become negative. Any tasks that were blocked in await will be woken if the count becomes <= 0.

CountDownLatches.get_count — Method

get_count(latch::CountDownLatch)

Returns the current count of the latch. This count may be negative.